fix: use f32 for position/rectangle

+++ livre-data/src/position.rs
@@ -4,8 +4,8 @@ use nom::sequence::separated_pair;
 
 #[derive(Debug, PartialEq, Clone, Copy)]
 pub struct Position {
-    pub x: i32,
-    pub y: i32,
+    pub x: f32,
+    pub y: f32,
 }
 
 impl Extract<'_> for Position {
@@ -16,7 +16,7 @@ impl Extract<'_> for Position {
 }
 
 impl Position {
-    pub fn new(x: i32, y: i32) -> Self {
+    pub fn new(x: f32, y: f32) -> Self {
         Self { x, y }
     }
 }
@@ -28,8 +28,9 @@ mod tests {
     use super::*;
 
     #[rstest]
-    #[case(b"0 0", Position::new(0, 0))]
-    #[case(b"-10 0", Position::new(-10, 0))]
+    #[case(b"0 0", Position::new(0.0, 0.0))]
+    #[case(b"-10 0", Position::new(-10.0, 0.0))]
+    #[case(b"-10.5 0.5", Position::new(-10.5, 0.5))]
     fn point(#[case] input: &[u8], #[case] expected: Position) {
         let (_, point) = extract(input).unwrap();
         assert_eq!(expected, point);
